Move nargo functions to not_nargo
