+++ src/not_nargo/mod.rs
@@ -55,3 +55,31 @@ fn lib_or_bin(current_path: &std::path::Path) -> Result<(std::path::PathBuf, Cra
         )),
     }
 }
+
+fn path_to_stdlib() -> std::path::PathBuf {
+    dirs::config_dir().unwrap().join("noir-lang").join("std/src")
+}
+
+fn add_std_lib(driver: &mut noirc_driver::Driver) {
+    let path_to_std_lib_file = path_to_stdlib().join("lib.nr");
+    let std_crate = driver.create_non_local_crate(path_to_std_lib_file, CrateType::Library);
+    let std_crate_name = "std";
+    driver.propagate_dep(std_crate, &noirc_frontend::graph::CrateName::new(std_crate_name).unwrap());
+}
+
+pub fn into_parsed_program<P: AsRef<std::path::Path>>(program_dir: P) -> (std::ffi::OsString, noirc_frontend::ParsedModule) {
+    let mut driver = Resolver::resolve_root_config(program_dir.as_ref(), &acvm::Language::R1CS).unwrap();
+    add_std_lib(&mut driver);
+    driver.build(true);
+
+    let mut errors = vec![];
+    // if driver.context.def_map(LOCAL_CRATE).is_some() {
+    //     return;
+    // }
+    let root_file_id = driver.context.crate_graph[noirc_frontend::graph::LOCAL_CRATE].root_file_id;
+    
+    let binding = std::path::PathBuf::from(driver.context.file_manager.as_simple_files().get(root_file_id.as_usize()).unwrap().name().to_string());
+    let file_name = binding.file_name().unwrap().to_os_string();
+
+    (file_name, noirc_frontend::hir::def_map::parse_file(&mut driver.context.file_manager, root_file_id, &mut errors))
+}
